@@ -1,22 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <UserSecretsId>034fbc91-b434-41de-9fb1-b4d29d91a6ac</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.13" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.1.13" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.13" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.13">
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="5.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.*">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.13" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.13" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.1.13" />
-    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="3.1.13" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="5.*" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="5.*" />
+    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="5.*" />
   </ItemGroup>
 
   <ItemGroup>
