@@ -27,8 +27,7 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.Payment
                 .PaymentLink.StoreValueAsUri(out var paymentLink)
                 .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.CreateReversal)].Should.BeVisible()
                 .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.PaidPayment)].Should.BeVisible()
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.ViewPayment)].Should.BeVisible()
-                .Actions.Rows.Count.Should.Equal(3);
+                .Actions.Rows[y => y.Name.Value.Contains(PaymentResourceOperations.ViewPayment)].Should.BeVisible();
 
             var swishPayment = await SwedbankPayClient.Payments.SwishPayments.Get(paymentLink, PaymentExpand.All);
 
@@ -49,34 +48,33 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.Payment
                         Is.EqualTo(State.Completed));
         }
 
-        [Test]
-        [Retry(3)]
-        [TestCaseSource(nameof(TestData), new object[] { false, PaymentMethods.Trustly })]
-        public async Task Payment_Trustly_Authorization(Product[] products, PayexInfo payexInfo)
-        {
-            GoToOrdersPage(products, payexInfo, Checkout.Option.Standard)
-                .PaymentLink.StoreValueAsUri(out var paymentLink)
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderReversal)].Should.BeVisible()
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
-                .Actions.Rows.Count.Should.Equal(2);
+        //[Test]
+        //[Retry(3)]
+        //[TestCaseSource(nameof(TestData), new object[] { false, PaymentMethods.Trustly })]
+        //public async Task Payment_Trustly_Authorization(Product[] products, PayexInfo payexInfo)
+        //{
+        //    GoToOrdersPage(products, payexInfo, Checkout.Option.Standard)
+        //        .PaymentLink.StoreValueAsUri(out var paymentLink)
+        //        .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderReversal)].Should.BeVisible()
+        //        .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible();
 
-            var trustlyPayment = await SwedbankPayClient.Payments.TrustlyPayments.Get(paymentLink, PaymentExpand.All);
+        //    var trustlyPayment = await SwedbankPayClient.Payments.TrustlyPayments.Get(paymentLink, PaymentExpand.All);
 
-            // Global Order
-            Assert.That(trustlyPayment.Payment.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
-            Assert.That(trustlyPayment.Payment.Currency.ToString(), Is.EqualTo("SEK"));
-            Assert.That(trustlyPayment.Payment.State, Is.EqualTo(State.Ready));
+        //    // Global Order
+        //    Assert.That(trustlyPayment.Payment.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
+        //    Assert.That(trustlyPayment.Payment.Currency.ToString(), Is.EqualTo("SEK"));
+        //    Assert.That(trustlyPayment.Payment.State, Is.EqualTo(State.Ready));
 
-            // Operations
-            Assert.That(trustlyPayment.Operations[LinkRelation.CreateReversal], Is.Not.Null);
-            Assert.That(trustlyPayment.Operations[LinkRelation.CreateCancellation], Is.Null);
-            Assert.That(trustlyPayment.Operations[LinkRelation.CreateCapture], Is.Null);
-            Assert.That(trustlyPayment.Operations[LinkRelation.PaidPayment], Is.Not.Null);
+        //    // Operations
+        //    Assert.That(trustlyPayment.Operations[LinkRelation.CreateReversal], Is.Not.Null);
+        //    Assert.That(trustlyPayment.Operations[LinkRelation.CreateCancellation], Is.Null);
+        //    Assert.That(trustlyPayment.Operations[LinkRelation.CreateCapture], Is.Null);
+        //    Assert.That(trustlyPayment.Operations[LinkRelation.PaidPayment], Is.Not.Null);
 
-            // Transactions
-            Assert.That(trustlyPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(1));
-            Assert.That(trustlyPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Sale).State,
-                        Is.EqualTo(State.Completed));
-        }
+        //    // Transactions
+        //    Assert.That(trustlyPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(1));
+        //    Assert.That(trustlyPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Sale).State,
+        //                Is.EqualTo(State.Completed));
+        //}
     }
 }
