@@ -21,87 +21,90 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Standard
         [Test]
         [Retry(3)]
         [TestCaseSource(nameof(TestData), new object[] { false, PaymentMethods.Card })]
-        public async Task Standard_PaymentOrder_Card_Authorization(Product[] products, PayexInfo payexInfo)
+        public void Standard_PaymentOrder_Card_Authorization(Product[] products, PayexInfo payexInfo)
         {
-            GoToOrdersPage(products, payexInfo, Checkout.Option.Standard)
+            Assert.DoesNotThrowAsync(async () => {
+
+                GoToOrdersPage(products, payexInfo, Checkout.Option.Standard)
                 .PaymentOrderLink.StoreValueAsUri(out var orderLink)
                 .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCancel)].Should.BeVisible()
                 .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCapture)].Should.BeVisible()
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
-                .Actions.Rows.Count.Should.Equal(3);
+                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible();
 
-            var order = await SwedbankPayClient.PaymentOrders.Get(orderLink, PaymentOrderExpand.All);
+                var order = await SwedbankPayClient.PaymentOrders.Get(orderLink, PaymentOrderExpand.All);
 
-            // Global Order
-            Assert.That(order.PaymentOrder.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
-            Assert.That(order.PaymentOrder.Currency.ToString(), Is.EqualTo("SEK"));
-            Assert.That(order.PaymentOrder.State, Is.EqualTo(State.Ready));
+                // Global Order
+                Assert.That(order.PaymentOrder.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
+                Assert.That(order.PaymentOrder.Currency.ToString(), Is.EqualTo("SEK"));
+                Assert.That(order.PaymentOrder.State, Is.EqualTo(State.Ready));
 
-            // Operations
-            Assert.That(order.Operations[LinkRelation.CreatePaymentOrderReversal], Is.Null);
-            Assert.That(order.Operations[LinkRelation.CreateCancellation], Is.Not.Null);
-            Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCapture], Is.Not.Null);
-            Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
+                // Operations
+                Assert.That(order.Operations[LinkRelation.CreatePaymentOrderReversal], Is.Null);
+                Assert.That(order.Operations[LinkRelation.CreateCancellation], Is.Not.Null);
+                Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCapture], Is.Not.Null);
+                Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
 
-            // Transactions
-            Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(1));
-            Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
-                        Is.EqualTo(State.Completed));
+                // Transactions
+                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(1));
+                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
+                            Is.EqualTo(State.Completed));
 
-            // Order Items
-            Assert.That(order.PaymentOrder.OrderItems.OrderItemList.Count, Is.EqualTo(products.Count()));
-            for (var i = 0; i < products.Count(); i++)
-            {
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Name, Is.EqualTo(products[i].Name));
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).UnitPrice.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice));
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Quantity, Is.EqualTo(products[i].Quantity));
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Amount.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice * products[i].Quantity));
-            }
+                // Order Items
+                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.Count, Is.EqualTo(products.Count()));
+                for (var i = 0; i < products.Count(); i++)
+                {
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Name, Is.EqualTo(products[i].Name));
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).UnitPrice.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice));
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Quantity, Is.EqualTo(products[i].Quantity));
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Amount.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice * products[i].Quantity));
+                }
+            });
         }
 
 
         [Test]
         [Retry(5)]
         [TestCaseSource(nameof(TestData), new object[] { false, PaymentMethods.Invoice })]
-        public async Task Standard_PaymentOrder_Invoice_Authorization(Product[] products, PayexInfo payexInfo)
+        public void Standard_PaymentOrder_Invoice_Authorization(Product[] products, PayexInfo payexInfo)
         {
-            GoToOrdersPage(products, payexInfo, Checkout.Option.Standard)
+            Assert.DoesNotThrowAsync(async () => {
+
+                GoToOrdersPage(products, payexInfo, Checkout.Option.Standard)
                 .PaymentOrderLink.StoreValueAsUri(out var orderLink)
                 .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCancel)].Should.BeVisible()
                 .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCapture)].Should.BeVisible()
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
-                .Actions.Rows.Count.Should.Equal(3);
+                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible();
 
-            var order = await SwedbankPayClient.PaymentOrders.Get(orderLink, PaymentOrderExpand.All);
+                var order = await SwedbankPayClient.PaymentOrders.Get(orderLink, PaymentOrderExpand.All);
 
-            // Global Order
-            Assert.That(order.PaymentOrder.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
-            Assert.That(order.PaymentOrder.Currency.ToString(), Is.EqualTo("SEK"));
-            Assert.That(order.PaymentOrder.State, Is.EqualTo(State.Ready));
+                // Global Order
+                Assert.That(order.PaymentOrder.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
+                Assert.That(order.PaymentOrder.Currency.ToString(), Is.EqualTo("SEK"));
+                Assert.That(order.PaymentOrder.State, Is.EqualTo(State.Ready));
 
-            // Operations
-            Assert.That(order.Operations[LinkRelation.CreatePaymentOrderReversal], Is.Null);
-            Assert.That(order.Operations[LinkRelation.CreateCancellation], Is.Not.Null);
-            Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCapture], Is.Not.Null);
-            Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
+                // Operations
+                Assert.That(order.Operations[LinkRelation.CreatePaymentOrderReversal], Is.Null);
+                Assert.That(order.Operations[LinkRelation.CreateCancellation], Is.Not.Null);
+                Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCapture], Is.Not.Null);
+                Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
 
-            // Transactions
-            Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(2));
-            Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Initialization).State,
-                        Is.EqualTo(State.Completed));
-            Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
-                        Is.EqualTo(State.Completed));
+                // Transactions
+                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(2));
+                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Initialization).State,
+                            Is.EqualTo(State.Completed));
+                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
+                            Is.EqualTo(State.Completed));
 
-            // Order Items
-            Assert.That(order.PaymentOrder.OrderItems.OrderItemList.Count, Is.EqualTo(products.Count()));
-            for (var i = 0; i < products.Count(); i++)
-            {
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Name, Is.EqualTo(products[i].Name));
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).UnitPrice.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice));
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Quantity, Is.EqualTo(products[i].Quantity));
-                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Amount.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice * products[i].Quantity));
-            }
+                // Order Items
+                Assert.That(order.PaymentOrder.OrderItems.OrderItemList.Count, Is.EqualTo(products.Count()));
+                for (var i = 0; i < products.Count(); i++)
+                {
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Name, Is.EqualTo(products[i].Name));
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).UnitPrice.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice));
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Quantity, Is.EqualTo(products[i].Quantity));
+                    Assert.That(order.PaymentOrder.OrderItems.OrderItemList.ElementAt(i).Amount.InLowestMonetaryUnit, Is.EqualTo(products[i].UnitPrice * products[i].Quantity));
+                }
+            });
         }
-
     }
 }
