@@ -22,21 +22,23 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Update
         [Test]
         [Retry(3)]
         [TestCaseSource(nameof(TestData), new object[] { false, PaymentMethods.Card })]
-        public async Task Update_PaymentOrder(Product[] products, PayexInfo payexInfo)
+        public void Update_PaymentOrder(Product[] products, PayexInfo payexInfo)
         {
-            GoToPayexSwishPaymentFrame(products.Skip(1).ToArray())
+            Assert.DoesNotThrowAsync( async () => {
+
+                GoToPayexSwishPaymentFrame(products.Skip(1).ToArray())
                 .Pay.Content.Should.BeEquivalent($"Betala {string.Format("{0:N2}", Convert.ToDecimal(products.Skip(1).First().UnitPrice / 100 * products.Skip(1).First().Quantity))} kr")
                 .SwitchToRoot<HomePage>()
                 .Header.Products.ClickAndGo();
 
-            GoToOrdersPage(products, payexInfo, Checkout.Option.Anonymous)
-                .PaymentOrderLink.StoreValueAsUri(out var orderLink)
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCancel)].Should.BeVisible()
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCapture)].Should.BeVisible()
-                .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
-                .Actions.Rows.Count.Should.Equal(3);
+                GoToOrdersPage(products, payexInfo, Checkout.Option.Anonymous)
+                    .PaymentOrderLink.StoreValueAsUri(out var orderLink)
+                    .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCancel)].Should.BeVisible()
+                    .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCapture)].Should.BeVisible()
+                    .Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible();
 
-            await SwedbankPayClient.PaymentOrders.Get(orderLink, PaymentOrderExpand.All);
+                await SwedbankPayClient.PaymentOrders.Get(orderLink, PaymentOrderExpand.All);
+            });
         }
     }
 }
